requirements.txt
pymongo==4.2.0
fastapi==0.79.0
uvicorn==0.18.2
bcrypt==4.0.0
PyJWT==2.5.0